Make ReportBreakageCategorySingleChoiceActivity activity scoped
