+++ app-tracking-protection/vpn-impl/src/main/java/com/duckduckgo/mobile/android/vpn/breakage/ReportBreakageMetadataReporter.kt
@@ -17,7 +17,7 @@
 package com.duckduckgo.mobile.android.vpn.breakage
 
 import com.duckduckgo.common.utils.DispatcherProvider
-import com.duckduckgo.di.scopes.VpnScope
+import com.duckduckgo.di.scopes.ActivityScope
 import com.duckduckgo.mobile.android.vpn.state.VpnStateCollector
 import com.squareup.anvil.annotations.ContributesBinding
 import javax.inject.Inject
@@ -27,7 +27,7 @@ interface ReportBreakageMetadataReporter {
     suspend fun getVpnStateMetadata(appPackageId: String? = null): String
 }
 
-@ContributesBinding(VpnScope::class)
+@ContributesBinding(ActivityScope::class)
 class RealReportBreakageMetadataReporter @Inject constructor(
     private val vpnStateCollector: VpnStateCollector,
     private val dispatcherProvider: DispatcherProvider,
