+++ app-tracking-protection/vpn-impl/src/main/java/com/duckduckgo/mobile/android/vpn/bugreport/VpnStateCollector.kt
@@ -19,7 +19,7 @@ package com.duckduckgo.mobile.android.vpn.bugreport
 import com.duckduckgo.anvil.annotations.ContributesPluginPoint
 import com.duckduckgo.common.utils.DispatcherProvider
 import com.duckduckgo.common.utils.plugins.PluginPoint
-import com.duckduckgo.di.scopes.VpnScope
+import com.duckduckgo.di.scopes.ActivityScope
 import com.duckduckgo.mobile.android.vpn.state.VpnStateCollector
 import com.duckduckgo.mobile.android.vpn.state.VpnStateCollectorPlugin
 import com.squareup.anvil.annotations.ContributesBinding
@@ -29,8 +29,8 @@ import kotlinx.coroutines.withContext
 import logcat.logcat
 import org.json.JSONObject
 
-@ContributesBinding(VpnScope::class)
-@SingleInstanceIn(VpnScope::class)
+@ContributesBinding(ActivityScope::class)
+@SingleInstanceIn(ActivityScope::class)
 class RealVpnStateCollector @Inject constructor(
     private val vpnStateCollectors: PluginPoint<VpnStateCollectorPlugin>,
     private val dispatcherProvider: DispatcherProvider,
@@ -51,7 +51,7 @@ class RealVpnStateCollector @Inject constructor(
 }
 
 @ContributesPluginPoint(
-    scope = VpnScope::class,
+    scope = ActivityScope::class,
     boundType = VpnStateCollectorPlugin::class,
 )
 @Suppress("unused")
