+++ app-tracking-protection/vpn-internal/src/main/java/com/duckduckgo/vpn/internal/feature/bugreport/VpnBugReporter.kt
@@ -17,7 +17,7 @@
 package com.duckduckgo.vpn.internal.feature.bugreport
 
 import com.duckduckgo.common.utils.DispatcherProvider
-import com.duckduckgo.di.scopes.VpnScope
+import com.duckduckgo.di.scopes.ActivityScope
 import com.duckduckgo.mobile.android.vpn.state.VpnStateCollector
 import com.squareup.anvil.annotations.ContributesBinding
 import javax.inject.Inject
@@ -28,7 +28,7 @@ interface VpnBugReporter {
     suspend fun generateBugReport(): String
 }
 
-@ContributesBinding(VpnScope::class)
+@ContributesBinding(ActivityScope::class)
 class RealVpnBugReporter @Inject constructor(
     private val vpnStateCollector: VpnStateCollector,
     private val dispatcherProvider: DispatcherProvider,
